refactor: rename Environment to EventLoop and move to event_loop module
